Avoid strcpy/memcpy since they didn't work(!)
Not sure what was happening or why, but the obs_mode string was not
getting properly retrieved from status memory.  It was tracked down to a
strcpy call that was not working for some unknown reason.  Replacing the
call to strcpy with a for loop that copies the characters one by one
worked.  Go figure...