Make gbtstatus_loop robust to "unknown" values
On 2016/11/13 (and reportedly the day before as well) the GBT Status
database exhibited previously unseen behavior of having the value
`unknown` for the `freq` field.  This raised an uncaught exception in
`guppi_utils.py` when attempting to convert this value to floating point
using `float()`.  The uncaught exception caused the
guppi_gbtstatus_loop.py script to terminate with an error.  This led to
key fields in the status buffer (e.g. source name and ra and dec and
obsfreq) not being updated, which led to mislabled datafiles.