Improve robustness of guppi_gbtstatus_loop.py
This involves holding the lock on the database from before reading
shared memory to after writing back to shared memory.  To prevent
holding the shared memory lock during connection to the GBT status
database, the connection to the database is now done before the main
loop in guppi_gbtstatus_loop.py.