Update SCAN in databuf header before marking full
Due to latencies in the system, sometimes the SCAN value is not updated
to the new value when the first packet of the scan arrives.  This SCAN
number is important because it ends up in the output filenames, so to
give it more time to propagate into the status buffer, we update the
SCAN value in the databuf header (copied from the status buffer when the
data block's first packet arrives) just before we mark the block as
filled (i.e. just before we give the block to the downstream thread,
e.g.  the disk thread).